wrote tests for callbackUtils
