+++ lib/routes/callback.js
@@ -1,18 +1,9 @@
 const { Router } = require('express');
 const getAccessToken = require('../services/getAccessToken');
-
-const attachOAuthToken = (res, token) => {
-  res.cookie('oauth-token', token, {
-    maxAge: 24 * 60 * 60 * 1000
-  });
-};
-
-const pullOutAccessToken = (accessTokenData) => {
-  const firstPass = accessTokenData.split('=');
-  const secondItem = firstPass[1];
-  const secondPass = secondItem.split('&');
-  return secondPass[0];
-}
+const {
+  attachOAuthToken,
+  pullOutAccessToken
+} = require('./callbackUtils');
 
 module.exports = Router()
   .get('/', async(req, res) => {
@@ -21,5 +12,4 @@ module.exports = Router()
     attachOAuthToken(res, accessToken);
     res.send(accessTokenData.text);
   });
-
